requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
pathlib2==2.3.7
tqdm==4.66.1
colorama==0.4.6